Move trim functions to datafusion-functions
