+++ datafusion/functions/Cargo.toml
@@ -29,10 +29,9 @@ authors = { workspace = true }
 rust-version = { workspace = true }
 
 [features]
-# enable string functions
-string_expressions = []
 # enable core functions
 core_expressions = []
+crypto_expressions = ["md-5", "sha2", "blake2", "blake3"]
 # enable datetime functions
 datetime_expressions = []
 # Enable encoding by default so the doctests work. In general don't automatically enable all packages.
@@ -51,7 +50,9 @@ encoding_expressions = ["base64", "hex"]
 math_expressions = []
 # enable regular expressions
 regex_expressions = ["regex"]
-crypto_expressions = ["md-5", "sha2", "blake2", "blake3"]
+# enable string functions
+string_expressions = []
+
 [lib]
 name = "datafusion_functions"
 path = "src/lib.rs"
