+++ datafusion/functions/src/string/ltrim.rs
@@ -15,38 +15,48 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use crate::string::common::{handle, utf8_to_str_type};
-use arrow::datatypes::DataType;
-use datafusion_common::Result;
-use datafusion_expr::ColumnarValue;
-use datafusion_expr::{ScalarUDFImpl, Signature, Volatility};
+use arrow::array::{ArrayRef, OffsetSizeTrait};
 use std::any::Any;
 
+use arrow::datatypes::DataType;
+
+use datafusion_common::{exec_err, Result};
+use datafusion_expr::TypeSignature::*;
+use datafusion_expr::{ColumnarValue, Volatility};
+use datafusion_expr::{ScalarUDFImpl, Signature};
+
+use crate::string::common::*;
+
+/// Returns the longest string  with leading characters removed. If the characters are not specified, whitespace is removed.
+/// ltrim('zzzytest', 'xyz') = 'test'
+fn ltrim<T: OffsetSizeTrait>(args: &[ArrayRef]) -> Result<ArrayRef> {
+    general_trim::<T>(args, TrimType::Left)
+}
+
 #[derive(Debug)]
-pub(super) struct UpperFunc {
+pub(super) struct LtrimFunc {
     signature: Signature,
 }
 
-impl UpperFunc {
+impl LtrimFunc {
     pub fn new() -> Self {
         use DataType::*;
         Self {
-            signature: Signature::uniform(
-                1,
-                vec![Utf8, LargeUtf8],
+            signature: Signature::one_of(
+                vec![Exact(vec![Utf8]), Exact(vec![Utf8, Utf8])],
                 Volatility::Immutable,
             ),
         }
     }
 }
 
-impl ScalarUDFImpl for UpperFunc {
+impl ScalarUDFImpl for LtrimFunc {
     fn as_any(&self) -> &dyn Any {
         self
     }
 
     fn name(&self) -> &str {
-        "upper"
+        "ltrim"
     }
 
     fn signature(&self) -> &Signature {
@@ -54,10 +64,14 @@ impl ScalarUDFImpl for UpperFunc {
     }
 
     fn return_type(&self, arg_types: &[DataType]) -> Result<DataType> {
-        utf8_to_str_type(&arg_types[0], "upper")
+        utf8_to_str_type(&arg_types[0], "ltrim")
     }
 
     fn invoke(&self, args: &[ColumnarValue]) -> Result<ColumnarValue> {
-        handle(args, |string| string.to_uppercase(), "upper")
+        match args[0].data_type() {
+            DataType::Utf8 => make_scalar_function(ltrim::<i32>, vec![])(args),
+            DataType::LargeUtf8 => make_scalar_function(ltrim::<i64>, vec![])(args),
+            other => exec_err!("Unsupported data type {other:?} for function ltrim"),
+        }
     }
 }
