+++ datafusion/functions/src/string/rtrim.rs
@@ -16,30 +16,29 @@
 // under the License.
 
 use arrow::array::{ArrayRef, OffsetSizeTrait};
-use arrow::datatypes::DataType;
-use datafusion_common::exec_err;
-use datafusion_common::Result;
-use datafusion_expr::ColumnarValue;
-use datafusion_expr::TypeSignature::*;
-use datafusion_expr::{ScalarUDFImpl, Signature, Volatility};
 use std::any::Any;
 
-use crate::string::{make_scalar_function, utf8_to_str_type};
+use arrow::datatypes::DataType;
 
-use super::{general_trim, TrimType};
+use datafusion_common::{exec_err, Result};
+use datafusion_expr::TypeSignature::*;
+use datafusion_expr::{ColumnarValue, Volatility};
+use datafusion_expr::{ScalarUDFImpl, Signature};
 
-/// Returns the longest string  with leading and trailing characters removed. If the characters are not specified, whitespace is removed.
-/// btrim('xyxtrimyyx', 'xyz') = 'trim'
-pub fn btrim<T: OffsetSizeTrait>(args: &[ArrayRef]) -> Result<ArrayRef> {
-    general_trim::<T>(args, TrimType::Both)
+use crate::string::common::*;
+
+/// Returns the longest string  with trailing characters removed. If the characters are not specified, whitespace is removed.
+/// rtrim('testxxzx', 'xyz') = 'test'
+fn rtrim<T: OffsetSizeTrait>(args: &[ArrayRef]) -> Result<ArrayRef> {
+    general_trim::<T>(args, TrimType::Right)
 }
 
 #[derive(Debug)]
-pub(super) struct TrimFunc {
+pub(super) struct RtrimFunc {
     signature: Signature,
 }
 
-impl TrimFunc {
+impl RtrimFunc {
     pub fn new() -> Self {
         use DataType::*;
         Self {
@@ -51,13 +50,13 @@ impl TrimFunc {
     }
 }
 
-impl ScalarUDFImpl for TrimFunc {
+impl ScalarUDFImpl for RtrimFunc {
     fn as_any(&self) -> &dyn Any {
         self
     }
 
     fn name(&self) -> &str {
-        "trim"
+        "rtrim"
     }
 
     fn signature(&self) -> &Signature {
@@ -65,14 +64,14 @@ impl ScalarUDFImpl for TrimFunc {
     }
 
     fn return_type(&self, arg_types: &[DataType]) -> Result<DataType> {
-        utf8_to_str_type(&arg_types[0], "trim")
+        utf8_to_str_type(&arg_types[0], "rtrim")
     }
 
     fn invoke(&self, args: &[ColumnarValue]) -> Result<ColumnarValue> {
         match args[0].data_type() {
-            DataType::Utf8 => make_scalar_function(btrim::<i32>, vec![])(args),
-            DataType::LargeUtf8 => make_scalar_function(btrim::<i64>, vec![])(args),
-            other => exec_err!("Unsupported data type {other:?} for function trim"),
+            DataType::Utf8 => make_scalar_function(rtrim::<i32>, vec![])(args),
+            DataType::LargeUtf8 => make_scalar_function(rtrim::<i64>, vec![])(args),
+            other => exec_err!("Unsupported data type {other:?} for function rtrim"),
         }
     }
 }
